refactor(servlet): cleanse package structure
