+++ tomcat/src/main/java/nextstep/jwp/service/UserService.java
@@ -4,8 +4,8 @@ import nextstep.Application;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.support.HttpException;
-import org.apache.coyote.response.HttpResponse;
-import org.apache.coyote.response.HttpResponse.HttpResponseBuilder;
+import org.apache.coyote.servlet.response.HttpResponse;
+import org.apache.coyote.servlet.response.HttpResponse.HttpResponseBuilder;
 import org.apache.coyote.support.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
