+++ tomcat/src/main/java/nextstep/jwp/servlet/CustomServlet.java
@@ -2,8 +2,8 @@ package nextstep.jwp.servlet;
 
 import org.apache.coyote.servlet.Servlet;
 import org.apache.coyote.support.HttpException;
-import org.apache.coyote.request.HttpRequest;
-import org.apache.coyote.response.HttpResponse;
+import org.apache.coyote.servlet.request.HttpRequest;
+import org.apache.coyote.servlet.response.HttpResponse;
 
 public class CustomServlet implements Servlet {
