+++ tomcat/src/main/java/nextstep/jwp/support/ResourceRegistry.java
@@ -4,9 +4,8 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import nextstep.jwp.support.ExceptionPage;
-import org.apache.coyote.response.HttpResponse;
-import org.apache.coyote.response.HttpResponse.HttpResponseBuilder;
+import org.apache.coyote.servlet.response.HttpResponse;
+import org.apache.coyote.servlet.response.HttpResponse.HttpResponseBuilder;
 import org.apache.coyote.support.HttpException;
 import org.apache.coyote.support.HttpStatus;
