+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -6,10 +6,10 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.coyote.Processor;
-import org.apache.coyote.request.HttpRequest;
-import org.apache.coyote.request.RequestHeaders;
-import org.apache.coyote.request.StartLine;
-import org.apache.coyote.response.HttpResponse;
+import org.apache.coyote.servlet.request.HttpRequest;
+import org.apache.coyote.servlet.request.RequestHeaders;
+import org.apache.coyote.servlet.request.StartLine;
+import org.apache.coyote.servlet.response.HttpResponse;
 import org.apache.coyote.servlet.Servlet;
 import org.apache.coyote.support.HttpException;
 import org.slf4j.Logger;
