+++ tomcat/src/main/java/org/apache/coyote/servlet/Servlet.java
@@ -1,7 +1,7 @@
 package org.apache.coyote.servlet;
 
-import org.apache.coyote.request.HttpRequest;
-import org.apache.coyote.response.HttpResponse;
+import org.apache.coyote.servlet.request.HttpRequest;
+import org.apache.coyote.servlet.response.HttpResponse;
 
 public interface Servlet {
